+++ python/fledge/services/core/routes.py
@@ -31,7 +31,7 @@ from fledge.services.core.api import package_log
 from fledge.services.core.api.repos import configure as configure_repo
 from fledge.services.core.api.control_service import script_management
 from fledge.services.core.api.control_service import acl_management
-from fledge.services.core.api import proxy
+
 
 __author__ = "Ashish Jabble, Praveen Garg, Massimiliano Pinto, Amarendra K Sinha"
 __copyright__ = "Copyright (c) 2017-2018 OSIsoft, LLC"
@@ -247,9 +247,6 @@ def setup(app):
 
     # enable cors support
     enable_cors(app)
-
-    # Proxy Public
-    proxy.setup(app)
 
     # enable a live debugger (watcher) for requests, see https://github.com/aio-libs/aiohttp-debugtoolbar
     # this will neutralize error middleware
